+++ example/materialBall.js
@@ -25,6 +25,8 @@ const params = {
 		metalness: 0.8,
 		ior: 1.495,
 		transmission: 0.0,
+		attenuationColor: '#ffffff',
+		attenuationDistance: 0.5,
 		opacity: 1.0,
 		clearcoat: 0.0,
 		clearcoatRoughness: 0.0,
@@ -45,6 +47,8 @@ const params = {
 		roughness: 0.9,
 		metalness: 0.1,
 		transmission: 0.0,
+		attenuationColor: '#ffffff',
+		attenuationDistance: 0.5,
 		ior: 1.495,
 		opacity: 1.0,
 		clearcoat: 0.0,
@@ -63,17 +67,7 @@ const params = {
 		color: '#000000',
 		roughness: 0.1,
 		metalness: 0.05,
-		clearcoat: 0.0,
-		clearcoatRoughness: 0.0,
-		sheenColor: '#000000',
-		sheenRoughness: 0.0,
-		iridescence: 0.0,
-		iridescenceIOR: 1.5,
-		iridescenceThickness: 400,
-		specularColor: '#ffffff',
-		specularIntensity: 1.0,
 		matte: false,
-		castShadow: true,
 	},
 
 	multipleImportanceSampling: true,
@@ -104,7 +98,9 @@ if ( window.location.hash.includes( 'transmission' ) ) {
 	params.material1.roughness = 0.23;
 	params.material1.transmission = 1.0;
 	params.material1.color = '#ffffff';
+
 	params.bounces = 10;
+	params.tiles = 2;
 
 } else if ( window.location.hash.includes( 'iridescent' ) ) {
 
@@ -112,6 +108,26 @@ if ( window.location.hash.includes( 'transmission' ) ) {
 	params.material1.roughness = 0.25;
 	params.material1.metalness = 1.0;
 	params.material1.iridescence = 1.0;
+
+} else if ( window.location.hash.includes( 'acrylic' ) ) {
+
+	params.material1.color = '#ffffff';
+	params.material1.roughness = 0;
+	params.material1.metalness = 0;
+	params.material1.transmission = 1.0;
+	params.material1.attenuationDistance = 0.75;
+	params.material1.attenuationColor = '#2a6dc6';
+
+	params.material2.color = '#ffffff';
+	params.material2.roughness = 0.0;
+	params.material2.metalness = 0.975;
+
+	params.material3.color = '#999999';
+	params.material3.roughness = 0.2;
+	params.material3.metalness = 0.0;
+
+	params.bounces = 20;
+	params.tiles = 3;
 
 }
 
@@ -158,11 +174,13 @@ async function init() {
 	blitQuad = new FullScreenQuad( new THREE.MeshBasicMaterial( {
 		map: ptRenderer.target.texture,
 		blending: THREE.CustomBlending,
+		premultipliedAlpha: renderer.getContextAttributes().premultipliedAlpha,
 	} ) );
 
 	denoiseQuad = new FullScreenQuad( new DenoiseMaterial( {
 		map: ptRenderer.target.texture,
 		blending: THREE.CustomBlending,
+		premultipliedAlpha: renderer.getContextAttributes().premultipliedAlpha,
 	} ) );
 
 	controls = new OrbitControls( perspectiveCamera, renderer.domElement );
@@ -336,6 +354,8 @@ async function init() {
 	denoiseFolder.add( params, 'denoiseSigma', 0.01, 12.0 );
 	denoiseFolder.add( params, 'denoiseThreshold', 0.01, 1.0 );
 	denoiseFolder.add( params, 'denoiseKSigma', 0.0, 12.0 );
+	denoiseFolder.close();
+
 
 	const envFolder = gui.addFolder( 'Environment' );
 	envFolder.add( params, 'environmentIntensity', 0, 10 ).onChange( () => {
@@ -377,6 +397,7 @@ async function init() {
 		}
 
 	} );
+	envFolder.close();
 
 	const cameraFolder = gui.addFolder( 'Camera' );
 	cameraFolder.add( params, 'cameraProjection', [ 'Perspective', 'Orthographic', 'Equirectangular' ] ).onChange( v => {
@@ -402,6 +423,7 @@ async function init() {
 		reset();
 
 	} ).listen();
+	cameraFolder.close();
 
 	const matFolder1 = gui.addFolder( 'Shell Material' );
 	matFolder1.addColor( params.material1, 'color' ).onChange( reset );
@@ -411,6 +433,8 @@ async function init() {
 	matFolder1.add( params.material1, 'metalness', 0, 1 ).onChange( reset );
 	matFolder1.add( params.material1, 'opacity', 0, 1 ).onChange( reset );
 	matFolder1.add( params.material1, 'transmission', 0, 1 ).onChange( reset );
+	matFolder1.add( params.material1, 'attenuationDistance', 0.05, 2.0 ).onChange( reset );
+	matFolder1.addColor( params.material1, 'attenuationColor' ).onChange( reset );
 	matFolder1.add( params.material1, 'ior', 0.9, 3.0 ).onChange( reset );
 	matFolder1.add( params.material1, 'clearcoat', 0, 1 ).onChange( reset );
 	matFolder1.add( params.material1, 'clearcoatRoughness', 0, 1 ).onChange( reset );
@@ -433,6 +457,8 @@ async function init() {
 	matFolder2.add( params.material2, 'metalness', 0, 1 ).onChange( reset );
 	matFolder2.add( params.material2, 'opacity', 0, 1 ).onChange( reset );
 	matFolder2.add( params.material2, 'transmission', 0, 1 ).onChange( reset );
+	matFolder2.add( params.material2, 'attenuationDistance', 0.05, 2.0 ).onChange( reset );
+	matFolder2.addColor( params.material2, 'attenuationColor' ).onChange( reset );
 	matFolder2.add( params.material2, 'ior', 0.9, 3.0 ).onChange( reset );
 	matFolder2.add( params.material2, 'clearcoat', 0, 1 ).onChange( reset );
 	matFolder2.add( params.material2, 'clearcoatRoughness', 0, 1 ).onChange( reset );
@@ -451,17 +477,7 @@ async function init() {
 	matFolder3.addColor( params.material3, 'color' ).onChange( reset );
 	matFolder3.add( params.material3, 'roughness', 0, 1 ).onChange( reset );
 	matFolder3.add( params.material3, 'metalness', 0, 1 ).onChange( reset );
-	matFolder3.add( params.material3, 'clearcoat', 0, 1 ).onChange( reset );
-	matFolder3.add( params.material3, 'clearcoatRoughness', 0, 1 ).onChange( reset );
-	matFolder3.addColor( params.material3, 'sheenColor' ).onChange( reset );
-	matFolder3.add( params.material3, 'sheenRoughness', 0, 1 ).onChange( reset );
 	matFolder3.add( params.material3, 'matte' ).onChange( reset );
-	matFolder3.add( params.material3, 'castShadow' ).onChange( reset );
-	matFolder3.add( params.material3, 'iridescence', 0.0, 1.0 ).onChange( reset );
-	matFolder3.add( params.material3, 'iridescenceIOR', 0.1, 3.0 ).onChange( reset );
-	matFolder3.add( params.material3, 'iridescenceThickness', 0.0, 1200.0 ).onChange( reset );
-	matFolder3.addColor( params.material3, 'specularColor' ).onChange( reset );
-	matFolder3.add( params.material3, 'specularIntensity', 0.0, 1.0 ).onChange( reset );
 	matFolder3.close();
 
 	animate();
@@ -562,6 +578,8 @@ function animate() {
 	m1.metalness = params.material1.metalness;
 	m1.roughness = params.material1.roughness;
 	m1.transmission = params.material1.transmission;
+	m1.attenuationDistance = params.material1.attenuationDistance;
+	m1.attenuationColor.set( params.material1.attenuationColor );
 	m1.ior = params.material1.ior;
 	m1.opacity = params.material1.opacity;
 	m1.clearcoat = params.material1.clearcoat;
@@ -573,6 +591,7 @@ function animate() {
 	m1.iridescenceThicknessRange = [ 0, params.material1.iridescenceThickness ];
 	m1.specularColor.set( params.material1.specularColor ).convertSRGBToLinear();
 	m1.specularIntensity = params.material1.specularIntensity;
+	m1.transparent = m1.opacity < 1;
 
 	const m2 = materials[ 1 ];
 	m2.color.set( params.material2.color ).convertSRGBToLinear();
@@ -581,6 +600,8 @@ function animate() {
 	m2.metalness = params.material2.metalness;
 	m2.roughness = params.material2.roughness;
 	m2.transmission = params.material2.transmission;
+	m2.attenuationDistance = params.material2.attenuationDistance;
+	m2.attenuationColor.set( params.material2.attenuationColor );
 	m2.ior = params.material2.ior;
 	m2.opacity = params.material2.opacity;
 	m2.clearcoat = params.material2.clearcoat;
@@ -592,20 +613,13 @@ function animate() {
 	m2.iridescenceThicknessRange = [ 0, params.material2.iridescenceThickness ];
 	m2.specularColor.set( params.material2.specularColor ).convertSRGBToLinear();
 	m2.specularIntensity = params.material2.specularIntensity;
+	m2.transparent = m2.opacity < 1;
 
 	const m3 = materials[ 2 ];
 	m3.color.set( params.material3.color ).convertSRGBToLinear();
 	m3.metalness = params.material3.metalness;
 	m3.roughness = params.material3.roughness;
-	m3.clearcoat = params.material3.clearcoat;
-	m3.clearcoatRoughness = params.material3.clearcoatRoughness;
-	m3.sheenColor.set( params.material3.sheenColor ).convertSRGBToLinear();
-	m3.sheenRoughness = params.material3.sheenRoughness;
-	m3.iridescence = params.material3.iridescence;
-	m3.iridescenceIOR = params.material3.iridescenceIOR;
-	m3.iridescenceThicknessRange = [ 0, params.material3.iridescenceThickness ];
-	m3.specularColor.set( params.material3.specularColor ).convertSRGBToLinear();
-	m3.specularIntensity = params.material3.specularIntensity;
+	m3.transparent = m3.opacity < 1;
 
 	ptRenderer.material.materials.updateFrom( sceneInfo.materials, sceneInfo.textures );
 	ptRenderer.material.materials.setMatte( 0, params.material1.matte );
@@ -613,7 +627,6 @@ function animate() {
 	ptRenderer.material.materials.setMatte( 2, params.material3.matte );
 	ptRenderer.material.materials.setCastShadow( 0, params.material1.castShadow );
 	ptRenderer.material.materials.setCastShadow( 1, params.material2.castShadow );
-	ptRenderer.material.materials.setCastShadow( 2, params.material3.castShadow );
 
 	ptRenderer.material.filterGlossyFactor = params.filterGlossyFactor;
 	ptRenderer.material.environmentIntensity = params.environmentIntensity;