+++ src/materials/PhysicalPathTracingMaterial.js
@@ -273,7 +273,7 @@ export class PhysicalPathTracingMaterial extends MaterialBase {
 									|| useAlphaTest && albedo.a < alphaTest
 
 									// opacity
-									|| ! useAlphaTest && albedo.a < rand()
+									|| material.transparent && ! useAlphaTest && albedo.a < rand()
 								)
 							) {
 
@@ -281,10 +281,15 @@ export class PhysicalPathTracingMaterial extends MaterialBase {
 
 							}
 
-							// only attenuate on the way in
-							if ( isBelowSurface ) {
+							if ( side == 1.0 && isBelowSurface ) {
 
+								// only attenuate by surface color on the way in
 								color *= mix( vec3( 1.0 ), albedo.rgb, transmissionFactor );
+
+							} else if ( side == - 1.0 ) {
+
+								// attenuate by medium once we hit the opposite side of the model
+								color *= transmissionAttenuation( dist, material.attenuationColor, material.attenuationDistance );
 
 							}
 
@@ -575,7 +580,7 @@ export class PhysicalPathTracingMaterial extends MaterialBase {
 							|| useAlphaTest && albedo.a < alphaTest
 
 							// opacity
-							|| ! useAlphaTest && albedo.a < rand()
+							|| material.transparent && ! useAlphaTest && albedo.a < rand()
 						) {
 
 							vec3 point = rayOrigin + rayDirection * dist;
@@ -782,6 +787,8 @@ export class PhysicalPathTracingMaterial extends MaterialBase {
 						surfaceRec.iridescenceThickness = iridescenceThickness;
 						surfaceRec.specularColor = specularColor;
 						surfaceRec.specularIntensity = specularIntensity;
+						surfaceRec.attenuationColor = material.attenuationColor;
+						surfaceRec.attenuationDistance = material.attenuationDistance;
 
 						// apply perceptual roughness factor from gltf
 						// https://registry.khronos.org/glTF/specs/2.0/glTF-2.0.html#microfacet-surfaces
@@ -931,6 +938,13 @@ export class PhysicalPathTracingMaterial extends MaterialBase {
 						}
 
 						throughputColor *= sampleRec.color / sampleRec.pdf;
+
+						// attenuate the throughput color by the medium color
+						if ( side == - 1.0 ) {
+
+							throughputColor *= transmissionAttenuation( dist, surfaceRec.attenuationColor, surfaceRec.attenuationDistance );
+
+						}
 
 						// discard the sample if there are any NaNs
 						if ( any( isnan( throughputColor ) ) || any( isinf( throughputColor ) ) ) {