+++ src/utils/GeometryPreparationUtils.js
@@ -5,12 +5,23 @@ export function getGroupMaterialIndicesAttribute( geometry, materials, allMateri
 	const indexAttr = geometry.index;
 	const posAttr = geometry.attributes.position;
 	const vertCount = posAttr.count;
-	const materialArray = new Uint8Array( vertCount );
 	const totalCount = indexAttr ? indexAttr.count : vertCount;
 	let groups = geometry.groups;
 	if ( groups.length === 0 ) {
 
 		groups = [ { count: totalCount, start: 0, materialIndex: 0 } ];
+
+	}
+
+	// use an array with the minimum precision required to store all material id references.
+	let materialArray;
+	if ( allMaterials.length <= 255 ) {
+
+		materialArray = new Uint8Array( vertCount );
+
+	} else {
+
+		materialArray = new Uint16Array( vertCount );
 
 	}
 
@@ -104,7 +115,7 @@ export function setCommonAttributes( geometry, options ) {
 	if ( ! geometry.attributes.color && ( attributes && attributes.includes( 'color' ) ) ) {
 
 		const vertCount = geometry.attributes.position.count;
-		geometry.setAttribute( 'color', new BufferAttribute( new Float32Array( vertCount * 4 ), 4, true ) );
+		geometry.setAttribute( 'color', new BufferAttribute( new Float32Array( vertCount * 4 ), 4 ) );
 
 	}